Revisions
I don't know how to do much on this website, so this is my best guess. I didn't understand the getReverse() method and it kept endlessly trying to run for me so I rewrote it and got it to work. I also got the getReverseComplement() method working. It calls both other methods in one line and returns the reverse complement. It's all that I call in main() to make it work. I think all that's left is the rest of the JavaDocs, further testing, and cleaning it up.